[tests] all in one test script to better see the score.
